Including the Sessions helper module into the Application controller.
We could make an entirely new module for authentication, but the
Sessions controller already comes equipped with a module, namely,
SessionsHelper. Moreover, such helpers are automatically included in
Rails views, so all we need to do to use the Sessions helper functions
in controllers is to include the module into the Application controller

By default, all the helpers are available in the views but not in the
controllers. We need the methods from the Sessions helper in both
places, so we have to include it explicitly.