Add a before_create callback to create remember_token.
Note that we’ve hashed the remember token using SHA1, a hashing
algorithm much faster than the Bcrypt algorithm used to hash user
passwords in Section 6.3.1, which is important because (as we will see
in Section 8.2.2) for signed-in users it will be run on every page. SHA1
is less secure than Bcrypt, but in the present case it is more than
sufficient because the token being hashed is already a 16-digit random
string; the SHA1 hexdigest of such a string is essentially uncrackable.
(The call to to_s is to make sure we can handle nil tokens, which
shouldn’t happen in browsers but sometimes happens in tests.)